Completion of Local Outlier Factor Algorithm
This help to identify outliers at a local level and removes the global effect of other algorithms which may make it less accurate. There are 4 key concepts to understand
- K Distance
- Reachibility Distance
- Local Reachibility Density (LRD)
- Local Outlier Factor

ADVANTAGES OF LOF
A point will be considered as an outlier if it is at a small distance to the extremely dense cluster. The global approach may not consider that point as an outlier. But the LOF can effectively identify the local outliers.

DISADVANTAGES OF LOF
Since LOF is a ratio, it is tough to interpret. There is no specific threshold value above which a point is defined as an outlier. The identification of an outlier is dependent on the problem and the user.

 CONCLUSION
Local outlier factor (LOF) values identify an outlier based on the local neighborhood. It gives better results than the global approach to find outliers. Since there is no threshold value of LOF, the selection of a point as an outlier is user-dependent.